抽取父类和公共方法,添加动画效果的切换及手势识别,防盗功能的设置
抽取父类和公共方法,添加动画效果的切换及手势识别,防盗功能的设置

+++ src/com/itheima/mobilesafe/utils/Md5Utils.java
@@ -10,7 +10,7 @@ public class Md5Utils {
 			byte[] result = digest.digest(password.getBytes());
 			StringBuffer sb = new StringBuffer();
 			for(byte b : result){
-				int number = b & 0xff ;//加盐  - 3;//10进制
+				int number = (int)(b & 0xff) ;//加盐  - 3;//10进制
 				String str = Integer.toHexString(number);
 				if(str.length()==1){
 					sb.append("0");
